rating calculate

+++ sukejuru/sekujuru_app/models.py
@@ -37,7 +37,7 @@ class Anime(models.Model):
     time = models.TimeField()
     genre = models.ManyToManyField(Genre)
     season = models.ManyToManyField(Season)
-    rating = models.IntegerField(default=5)
+    rating = models.FloatField(default=5)
 
     def __str__(self):
         return f'{self.anime_id}: {self.anime_name} {self.time} {self.rating}'
